full-pipeline/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==1.3.1
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
coverage==7.9.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
humanfriendly==10.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.30.0
ipython==8.37.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpointer==2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.5.1
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
neurokit2==0.2.11
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==27.0.0
qudida==0.0.4
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
simsimd==6.4.9
six==1.17.0
sklearn-compat==0.1.3
soundfile==0.13.1
soupsieve==2.7
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.16
tinycss2==1.4.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.162
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wfdb==4.3.0
wget==3.2
wrapt==1.17.2
yarl==1.20.1
zenodo-get==2.0.0
